20211112 solving libft (cleated memory function)

+++ ft_strlen.c
@@ -6,14 +6,16 @@
 /*   By: chsong <chsong@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/11 00:08:50 by chsong            #+#    #+#             */
-/*   Updated: 2021/11/11 00:08:51 by chsong           ###   ########.fr       */
+/*   Updated: 2021/11/11 15:52:35 by chsong           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-unsigned long    ft_strlen(const char *s)
+#include "libft.h"
+
+size_t    ft_strlen(const char *s)
 {
     int n;
-
+    
     n = -1;
     while (s[++n])
         continue;
